--- Samples~/AvatarRenderSamples/MultipleRenders/MultipleAvatarRenderExample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2b358f16db09e64994ed07b04a9b298
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
