+++ Samples~/AvatarRenderSamples/MultipleRenders/RenderPanel.cs
@@ -1,4 +1,3 @@
-using System.Linq;
 using ReadyPlayerMe.Core;
 using UnityEngine;
 using UnityEngine.UI;
@@ -11,8 +10,6 @@ namespace ReadyPlayerMe.Samples
         [SerializeField] private AvatarRenderSettings renderSettings;
         [SerializeField] private Image image;
         
-        public Text heading;
-
         private void Start()
         {
             var avatarRenderLoader = new AvatarRenderLoader();
@@ -20,18 +17,11 @@ namespace ReadyPlayerMe.Samples
             avatarRenderLoader.LoadRender(url, renderSettings);
         }
 
-        public void SetHeading(string text)
-        {
-            var headingText = string.Concat(text.Select(x => char.IsUpper(x) ? " " + x : x.ToString())).TrimStart(' ');
-            heading.text = headingText;
-        }
-
-        public void SetImage(Texture2D texture)
+        private void SetImage(Texture2D texture)
         {
             var sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), new Vector2(.5f, .5f));
             image.sprite = sprite;
             image.preserveAspect = true;
         }
-
     }
 }
