--- Samples~/AvatarRenderSamples/SingleRender.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9880f04e6460c493d81cd7d9c46c9166
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
