--- Samples~/AvatarRenderSamples/SingleRender/AvatarRenderExample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62b7a236cf0534ae682c6aaa43737760
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
