--- Samples~/AvatarRenderSamples/SingleRender/AvatarRenderExample.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6a938d18afe2d4495be4025f1e367c15
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
